requirements.txt
asgiref==3.9.1
boto3==1.40.30
botocore==1.40.30
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
phonenumbers==9.0.13
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise